+++ react-client/src/layouts/TrackComparisonContainer.tsx
@@ -7,7 +7,7 @@ import { FaChevronDown } from 'react-icons/fa';
 import platform, { PlatformName } from '@constants/platform';
 
 import 'chart.js/auto';
-import { TrackComparisonLineChartWrapper } from '@layouts/TrackComparisonBoxWithLineChartWrapper';
+import { TrackComparisonWrapper } from '@layouts/TrackComparisonWrapper';
 import { SelectedTrack } from '@pages/ExplorePage';
 import { useCachedTracks } from '@hooks/useCachedTracks';
 import LoadingSpinner from '@components/LoadingSpinner';
@@ -120,7 +120,7 @@ export function TrackComparisonContainer({ selectedTracks }:Prob) {
 
         {/* contents */}
         <div className='p-2 '>
-          <TrackComparisonLineChartWrapper
+          <TrackComparisonWrapper
               selectedTracks={mappedSelectedTracks}
               selectedPlatformName={selectedPlatformName}
               startDate={startDate}
